+++ src/create-quiver/lib/json.js
@@ -1,0 +1,14 @@
+function stripJsonComments(text) {
+  return String(text || '')
+    .replace(/^\s*\/\/.*$/gm, '')
+    .replace(/\/\*[\s\S]*?\*\//g, '');
+}
+
+function parseJsonWithComments(text) {
+  return JSON.parse(stripJsonComments(text));
+}
+
+module.exports = {
+  parseJsonWithComments,
+  stripJsonComments,
+};